airflow/requirements.txt
apache-airflow-providers-amazon
pyarrow
pandas
sqlalchemy
psycopg2-binary
dbt-redshift
